ETH.Build simulation Issue
In the tutorial for gaining familiarity with Ethereum on ETH.Build, there are missing tabs such as account | Pool in this simulation Ethereum recommends. It's very difficult to follow the steps shown by @AustinGriffith on YouTube. 

Creating transactions between 'bob' and 'alice' in the object block cannot be completed in the simulation because the dummy account block is not in any of the tabs nor can you type it in from the draft board. 

Regarding 'mining' Ethereum, the 'pool' block and 'network' block are missing even though Austin is using both blocks in his 'Blockchain' ETH.BUILD tutorial simulation. 

Austin's YouTube tutorials were uploaded in 2019. Does anyone know if there is an upgrade to those tabs listed. 

Attached my ETH.Build progress | Creating a transaction between 'bob' and 'alice' using pre-formatted 'object' block. 

Thanks!